webapp/requirements.txt
flask
flask-caching
numpy
tensorflow-cpu
datetime
gunicorn
json-lines
pandas